Moved CroptimizR files into their own directory.
