+++ Models/Optimisation/Criteria.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Models.Sensitivity
+namespace Models.Optimisation
 {
     /// <summary>
     /// Criteria used to estimate parameters by minimizing the difference
