+++ Models/Optimisation/IOptimizationMethod.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Models.Sensitivity
+namespace Models.Optimisation
 {
     /// <summary>
     /// Encapsulates an optimization method which can be used by <see cref="CroptimizR"/>.
