+++ Models/Optimisation/Likelihood.cs
@@ -4,11 +4,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Models.Sensitivity
+namespace Models.Optimisation
 {
     /// <summary>
     /// Likelihood functions used as the crit function to estimate parameters
     /// using bayesian methods via CroptimizR.
+    /// 
+    /// https://sticsrpacks.github.io/CroptimizR/reference/Likelihoods.html
     /// </summary>
     public enum Likelihood
     {
